MDEV-37356 Annotate_rows written in a 'random' position
This is a review/optimization commit.
In order to not penalize with unnecessary table loop execution of
all-tables-transactional statement let's make up a special
branch for the mixed engine case.

The logics of the base (bbb34114ed0) commit remains just with no non-transactional tables present
`binlog_using_only_trans_tables` is `true` at once and that avoid the
loop where it's computed.

The 2nd loop is partly restored to its pre-base's (bbb34114ed0~1) form in that it
`restore`s `file->row_*` flags.
While combining two types of `file->row_*` restore the current shape
of the 2nd loop is not perfect so the following commit is going to
unify them, and by that to eliminate/optimize-away
a new TABLE::restore_row_logging (suggested by Serg) introduced by the base commit.

Mtr tests pass.