MDEV-36861  TPC-H Query 21 takes 10X longer on 11.4 compared to 10.11
commit b66cdbd1eaee changed the cost model from row based
to millisecond based.  Part of the commit touched
Duplicate_weedout_picker::check_qep() and subtly changed how
the cost calculation was done for the case where the first non-constant
table in the join is one converted into a semijoin.  Previous to this commit
the starting number of rows for determining the cost was
join->positions[first_tab].prefix_record_count.  After this commit
the number of rows was estimated to be 1.  This commit reverts that
change.

A number of test result files have reverted to the query plans prior to
the above commit.